requirements.txt
# AWS
boto3>=1.28.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.0

# Vector Store & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM
openai>=1.0.0

# Web Framework (optional API)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Document Generation
python-docx>=1.0.0
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
